# Conflicts:
#	CMakeLists.txt
#	src/monosat/simp/SimpSolver.h

+++ src/monosat/core/Optimize.cpp
@@ -33,8 +33,8 @@
 namespace Monosat{
 
 namespace Optimization{
-static long long time_limit=-1;
-static long long memory_limit=-1;
+static int64_t time_limit=-1;
+static int64_t memory_limit=-1;
 
 static bool has_system_time_limit=false;
 static rlim_t system_time_limit;
@@ -95,7 +95,7 @@ void enableResourceLimits(Solver * S){
 
 		if (time_limit < INT32_MAX && time_limit>=0) {
 			assert(cur_time>=0);
-			long long local_time_limit = 	 time_limit+cur_time;//make this a relative time limit
+			int64_t local_time_limit = 	 time_limit+cur_time;//make this a relative time limit
 			if (rl.rlim_max == RLIM_INFINITY || (rlim_t) local_time_limit < rl.rlim_max) {
 				rl.rlim_cur = local_time_limit;
 				if (setrlimit(RLIMIT_CPU, &rl) == -1)
@@ -232,7 +232,7 @@ int64_t optimize_linear_bv(Monosat::SimpSolver * S, Monosat::BVTheorySolver<int6
 	int64_t value = getApprox(bvTheory,bvID,!invert);
 	int64_t last_decision_value=value;
 	if(opt_verb>=1 || opt_verb_optimize>=1){
-		printf("Best bv%d = %ld",bvID,value);
+		printf("Best bv%d = %" PRId64 "",bvID,value);
 	}
 	// int bvID,const Weight & to, Var outerVar = var_Undef, bool decidable=true
 
@@ -272,16 +272,16 @@ int64_t optimize_linear_bv(Monosat::SimpSolver * S, Monosat::BVTheorySolver<int6
 		n_solves++;
 
 		if(opt_verb_optimize>=2){
-			printf("Testing bv%d %s %ld...\n",bvID, invert? ">=":"<=",value);
+			printf("Testing bv%d %s %" PRId64 "...\n",bvID, invert? ">=":"<=",value);
 		}
 
-		int conflict_limit = S->getConflictBudget();
+		int64_t conflict_limit = S->getConflictBudget();
 		if(conflict_limit<0)
 			conflict_limit=INT32_MAX;
-		int opt_lim = opt_limit_optimization_conflicts;
+        int64_t opt_lim = opt_limit_optimization_conflicts;
 		if(opt_lim<=0)
 			opt_lim=INT32_MAX;
-		int limit = std::min(opt_lim,conflict_limit);
+        int64_t limit = std::min(opt_lim,conflict_limit);
 		if(limit>= INT32_MAX){
 			limit=-1;//disable limit.
 		}
@@ -329,11 +329,11 @@ int64_t optimize_linear_bv(Monosat::SimpSolver * S, Monosat::BVTheorySolver<int6
 			}
 			value=value2;
 			if(opt_verb_optimize>=2){
-				printf("Found bv%d = %ld...\n",bvID,value);
+				printf("Found bv%d = %" PRId64 "...\n",bvID,value);
 			}
 			assume.pop();
 			if(opt_verb>=1){
-				printf("\rMin bv%d = %ld",bvID,value);
+				printf("\rMin bv%d = %" PRId64 "",bvID,value);
 			}
 		}else{
 			assume.pop();
@@ -410,7 +410,7 @@ int evalPB(SimpSolver & S,const Objective & o, bool over_approx, bool eval_at_le
 	}
 	return sum_val;
 }
-int64_t optimize_linear_pb(Monosat::SimpSolver * S, PB::PBConstraintSolver * pbSolver, bool invert, const vec<Lit> & assumes, const Objective & o , bool & hit_cutoff, int64_t & n_solves, bool & found_model){
+int optimize_linear_pb(Monosat::SimpSolver * S, PB::PBConstraintSolver * pbSolver, bool invert, const vec<Lit> & assumes, const Objective & o , bool & hit_cutoff, int64_t & n_solves, bool & found_model){
 	hit_cutoff=false;
 
 	vec<Lit> discarded_pb_constraints;
@@ -563,7 +563,7 @@ int64_t optimize_linear_pb(Monosat::SimpSolver * S, PB::PBConstraintSolver * pbS
 }
 
 
-int64_t optimize_binary_pb(Monosat::SimpSolver * S,  PB::PBConstraintSolver * pbSolver, bool invert, const vec<Lit> & assumes,const Objective & o,  bool & hit_cutoff, int64_t & n_solves, bool & found_model){
+int optimize_binary_pb(Monosat::SimpSolver * S,  PB::PBConstraintSolver * pbSolver, bool invert, const vec<Lit> & assumes,const Objective & o,  bool & hit_cutoff, int64_t & n_solves, bool & found_model){
 	hit_cutoff=false;
 
 	vec<Lit> discarded_pb_constraints;
@@ -598,10 +598,10 @@ int64_t optimize_binary_pb(Monosat::SimpSolver * S,  PB::PBConstraintSolver * pb
 		suggested_next_midpoint =	underapprox_sat_val;
 
 	bool first_round=true;
-	int64_t last_decision_value=max_val;
+    int last_decision_value=max_val;
 	//Lit last_decision_lit =  lit_Undef;//  pbSolver->addConditionalConstr(o.pb_lits, o.pb_weights,max_val, PB::Ineq::LEQ); //bvTheory->toSolver(bvTheory->newComparison(Comparison::leq,bvID,max_val,var_Undef,opt_decide_optimization_lits));
 	while(lt(min_val, max_val,invert) && !hit_cutoff){
-		int64_t mid_point = min_val + (max_val - min_val) / 2;
+		int mid_point = min_val + (max_val - min_val) / 2;
 
 		if (geq(mid_point, max_val,invert))
 			mid_point = invert ? max_val+1 : max_val - 1;//can this ever happen?
@@ -617,13 +617,13 @@ int64_t optimize_binary_pb(Monosat::SimpSolver * S,  PB::PBConstraintSolver * pb
 			assume.push(~l);
 		}
 		{
-			int conflict_limit = S->getConflictBudget();
+            int64_t conflict_limit = S->getConflictBudget();
 			if(conflict_limit<0)
 				conflict_limit=INT32_MAX;
-			int opt_lim = opt_limit_optimization_conflicts;
+            int64_t opt_lim = opt_limit_optimization_conflicts;
 			if(opt_lim<=0)
 				opt_lim=INT32_MAX;
-			int limit = std::min(opt_lim,conflict_limit);
+            int64_t limit = std::min(opt_lim,conflict_limit);
 			if(limit>= INT32_MAX){
 				limit=-1;//disable limit.
 			}
@@ -677,7 +677,7 @@ int64_t optimize_binary_pb(Monosat::SimpSolver * S,  PB::PBConstraintSolver * pb
 			if(geq(new_value,max_val,invert)){
 				throw std::runtime_error("Error in optimization (minimum values are inconsistent with model)");
 			}
-			int64_t underapprox_sat_val =  evalPB(*S, o, invert); //bvTheory->getUnderApprox(bvID);
+			int underapprox_sat_val =  evalPB(*S, o, invert); //bvTheory->getUnderApprox(bvID);
 			if(lt(underapprox_sat_val,max_val,invert))
 				suggested_next_midpoint =	underapprox_sat_val;
 			assert(leq(new_value,mid_point,invert));
@@ -783,7 +783,7 @@ int64_t optimize_binary_bv(Monosat::SimpSolver * S, Monosat::BVTheorySolver<int6
 	int64_t min_val = getApprox(bvTheory, bvID, invert,true); //bvTheory->getUnderApprox(bvID,true);
 	int64_t max_val = getApprox(bvTheory, bvID, !invert); //bvTheory->getOverApprox(bvID);
 	if(opt_verb>=1 || opt_verb_optimize>=1){
-		printf("Best bv%d = %ld",bvID,max_val);
+		printf("Best bv%d = %" PRId64 "",bvID,max_val);
 	}
 	int64_t suggested_next_midpoint = -1;
 
@@ -853,7 +853,7 @@ int64_t optimize_binary_bv(Monosat::SimpSolver * S, Monosat::BVTheorySolver<int6
 
 		n_solves++;
 		if(opt_verb_optimize>=2){
-			printf("Testing bv%d %s %ld...\n",bvID, invert? ">=":"<=",mid_point);
+			printf("Testing bv%d %s %" PRId64 "...\n",bvID, invert? ">=":"<=",mid_point);
 		}
 		{
 			int conflict_limit = S->getConflictBudget();
@@ -911,7 +911,7 @@ int64_t optimize_binary_bv(Monosat::SimpSolver * S, Monosat::BVTheorySolver<int6
 			assert(lt(new_value,max_val,invert));
 			max_val=new_value;
 			if(opt_verb_optimize>=2){
-				printf("Found bv%d = %ld...\n",bvID,max_val);
+				printf("Found bv%d = %" PRId64 "...\n",bvID,max_val);
 			}
 			if(leq(new_value,min_val,invert)){
 				//this can only happen if a budget was used and the solver quit early.
@@ -919,7 +919,7 @@ int64_t optimize_binary_bv(Monosat::SimpSolver * S, Monosat::BVTheorySolver<int6
                 assert(geq(min_val,getApprox(bvTheory,bvID,invert,true),invert)); //assert(min_val>=bvTheory->getUnderApprox(bvID,true));
 			}
 			if(opt_verb>=1 || opt_verb_optimize>=1){
-				printf("\rBest bv%d = %ld",bvID,max_val);
+				printf("\rBest bv%d = %" PRId64 "",bvID,max_val);
 			}
 		}else{
 			min_val = invert ? mid_point-1 : mid_point+1; //yes this is intentionally backward
@@ -1006,7 +1006,7 @@ int64_t optimize_binary_restart_bv(Monosat::SimpSolver * S, Monosat::BVTheorySol
 	int64_t min_val = getApprox(bvTheory, bvID, invert,true); //bvTheory->getUnderApprox(bvID,true);
 	int64_t max_val = getApprox(bvTheory, bvID, !invert); //bvTheory->getOverApprox(bvID);
 	if(opt_verb>=1 || opt_verb_optimize>=1){
-		printf("Best bv%d = %ld",bvID,max_val);
+		printf("Best bv%d = %" PRId64 "",bvID,max_val);
 	}
 	int64_t suggested_next_midpoint = -1;
 
@@ -1081,7 +1081,7 @@ int64_t optimize_binary_restart_bv(Monosat::SimpSolver * S, Monosat::BVTheorySol
 
 		n_solves++;
 		if(opt_verb_optimize>=2){
-			printf("Testing bv%d %s %ld...\n",bvID, invert? ">=":"<=",mid_point);
+			printf("Testing bv%d %s %" PRId64 "...\n",bvID, invert? ">=":"<=",mid_point);
 		}
 		{
 			int conflict_limit = S->getConflictBudget();
@@ -1131,7 +1131,7 @@ int64_t optimize_binary_restart_bv(Monosat::SimpSolver * S, Monosat::BVTheorySol
 			last_decision_value=mid_point;
 			int64_t new_value = getApprox(bvTheory, bvID, !invert); //bvTheory->getOverApprox(bvID);
 			if(opt_verb_optimize>=2){
-				printf("Found bv%d = %ld...\n",bvID,new_value);
+				printf("Found bv%d = %" PRId64 "...\n",bvID,new_value);
 			}
 			if(geq(new_value,max_val,invert)){
 				throw std::runtime_error("Error2 in optimization (minimum values are inconsistent with model)");
@@ -1149,7 +1149,7 @@ int64_t optimize_binary_restart_bv(Monosat::SimpSolver * S, Monosat::BVTheorySol
 				assert(gt(min_val,getApprox(bvTheory,bvID,invert,true),invert)); //assert(min_val>=bvTheory->getUnderApprox(bvID,true));
 			}
 			if(opt_verb>=1 || opt_verb_optimize>=1){
-				printf("\rBest bv%d = %ld",bvID,max_val);
+				printf("\rBest bv%d = %" PRId64 "",bvID,max_val);
 			}
 			suggested_next_midpoint=-1;
 			apparent_min_val=min_val;
@@ -1169,7 +1169,7 @@ int64_t optimize_binary_restart_bv(Monosat::SimpSolver * S, Monosat::BVTheorySol
 			if (gt(new_apparent_limit, apparent_min_val, invert) || ((drand(S->random_seed)<=opt_rnd_optimization_restart_freq))){
 				apparent_min_val = new_apparent_limit;
 				if(opt_verb_optimize>=2)
-					printf("Updating apparent min val to %ld.\n", apparent_min_val);
+					printf("Updating apparent min val to %" PRId64 ".\n", apparent_min_val);
 			}
 			if (!invert)
 				suggested_next_midpoint = apparent_min_val + (max_val - apparent_min_val) / 2;
@@ -1180,7 +1180,7 @@ int64_t optimize_binary_restart_bv(Monosat::SimpSolver * S, Monosat::BVTheorySol
 				}
 			}
 			if(opt_verb_optimize>=2)
-				printf("Suggested mid point is %ld (from minval %ld, maxval %ld, apparent_min_val %ld, current_apparent_min %ld).\n", suggested_next_midpoint,min_val,max_val, apparent_min_val, new_apparent_limit);
+				printf("Suggested mid point is %" PRId64 " (from minval %" PRId64 ", maxval %" PRId64 ", apparent_min_val %" PRId64 ", current_apparent_min %" PRId64 ").\n", suggested_next_midpoint,min_val,max_val, apparent_min_val, new_apparent_limit);
 			number_of_restarts++;
 
 			if(drand(S->random_seed)<=opt_rnd_optimization_freq){
@@ -1333,7 +1333,7 @@ lbool optimize_and_solve(SimpSolver & S,const vec<Lit> & assumes,const vec<Objec
 				if (objectives[i].isBV()) {
 					int bvID =objectives[i].bvID;
 					vec<Lit>  lits;
-					((BVTheorySolver<long>*) S.getBVTheory())->getLiterals(bvID, lits);
+					((BVTheorySolver<int64_t>*) S.getBVTheory())->getLiterals(bvID, lits);
 					for(Lit l:lits){
 						//printf("Setting priority for %d to %d\n", var(l), 100);
 						old_decision_priority.push();
@@ -1459,7 +1459,7 @@ lbool optimize_and_solve(SimpSolver & S,const vec<Lit> & assumes,const vec<Objec
 					}
 					assert(min_values[i] >= 0);
 					if (opt_verb >= 1 || opt_verb_optimize>=1) {
-						printf("\rBest bv%d = %ld\n", bvID, objectives[i].maximize? max_values[i]: min_values[i]);
+						printf("\rBest bv%d = %" PRId64 "\n", bvID, objectives[i].maximize? max_values[i]: min_values[i]);
 					}
 
 				}else{
@@ -1467,7 +1467,7 @@ lbool optimize_and_solve(SimpSolver & S,const vec<Lit> & assumes,const vec<Objec
 					if (opt_verb >= 1 || opt_verb_optimize>=1) {
 						printf("%s pb (%d of %d)\n",objectives[i].maximize ? "Maximizing": "Minimizing", i + 1, objectives.size());
 					}
-					int64_t val=0;
+					int val=0;
 					if (opt_optimization_search_type==1) {
 						val = optimize_linear_pb(&S, pbSolver,objectives[i].maximize, assume, objectives[i], hit_cutoff, n_solves, ever_solved);
 					} else if (opt_optimization_search_type==0) {
@@ -1580,12 +1580,12 @@ lbool optimize_and_solve(SimpSolver & S,const vec<Lit> & assumes,const vec<Objec
 			assert(r);
 
 			if(opt_verb>0 || opt_verb_optimize>=1){
-				printf("Best values found (after %ld calls) : ",n_solves);
+				printf("Best values found (after %" PRId64 " calls) : ",n_solves);
 				for(int i = 0;i<min_values.size();i++){
 					int64_t best_value = objectives[i].maximize ? max_values[i] : min_values[i];
 					if(objectives[i].isBV()) {
 						int bvID = objectives[i].bvID;
-						printf("bv%d=%ld,", bvID, best_value);
+						printf("bv%d=%" PRId64 ",", bvID, best_value);
 					}else{
 
 						printf("pb (");
@@ -1594,7 +1594,7 @@ lbool optimize_and_solve(SimpSolver & S,const vec<Lit> & assumes,const vec<Objec
 							int weight = objectives[i].pb_weights[j];
 							printf("%d * %dL +",weight,dimacs(l));
 						}
-						printf(") =%ld,", best_value);
+						printf(") =%" PRId64 ",", best_value);
 					}
 				}
 				printf("\n");
